+++ rf4s/component/tackle.py
@@ -331,7 +331,6 @@ class Tackle:
         while i > 0:
             i = utils.sleep_and_decrease(i, LOOP_DELAY)
             if self.detection.is_fish_captured():
-                self.landing_net_out = False
                 return
             if self.cfg.BOT.SNAG_DETECTION and self.detection.is_line_snagged():
                 raise exceptions.LineSnaggedError