+++ docs/en/CHANGELOG.md
@@ -1,5 +1,6 @@
 - Fix lure changing bug
 - Fix safe check bug
+- Fix telescopic mode landing net bug
 
 
 ## 0.7.3 (2025-09-10)